feat(result): result component rewrited to SuccessResult, FailedResult and Result components (module4-task1)
