+++ src/app/game-screen/failed-result/index.js
@@ -1,0 +1,4 @@
+import FailedResult from "./failed-result";
+
+
+export default FailedResult;
