+++ src/app/game-screen/success-result/index.js
@@ -1,0 +1,4 @@
+import SuccessResult from "./success-result";
+
+
+export default SuccessResult;
