+++ src/components/result/index.js
